introduce abstraction over keyboard events
In anticipation of support for backends other than evdev, it makes sense
at this early stage to start making use of an input event abstraction
layer that is specific to QSK. Some might say YAGNI or call it premature
optimization, but the use of evdev was really bothering me as I began
experimenting with different types of input transformations.

* introduce new `input.event` package with:
  * KeyboardEvent
  * KeyState
  * KeyCode
* update types in `device.linux` to expose KeyboardEvent types rather
than evdev_rs::InputEvent; this does significantly reduce the scope of
what we transform, but the proper focus for now is Keyboard events.
* implement a bunch of `std::convert::Into`s in order to seamlessly
handle conversions between `evdev_rs::InputEvent` and
`event::KeyboardEvent`.
* add `chrono` as a dependency to keep track of time values in
`event::KeyboardEvent`
* update to current latest commit on evdev-rs git repo