qsk: implement --log-to-file to log remap sessions
This is accomplished with the use of the new `recorder` module, which
provides `Listener` and `Recorder` types, which are connected by a
`async_std::channel::Sender`/`async_std::channel::Receiver` pair.

`Listener` implements `layers::InputTransformer` by wrapping a given
`layers::InputTransformer` and performing the following series of actions:

* cloning each `events::InputEvent` it receives and sending those via
  its `async_std::channel::Sender` to the paired `Recorder` (wrapped in
  a `layers::ControlCode::InputEvent` variant)
* passing the `events::InputEvent` it receives to the "inner"
  `layers::InputTransformer.transform` method
* cloning the `layers::ControlCode`s it receives from its "inner"
  `layers::InputTransformer.transform` method and sending those via its
  `async_std::channel::Receiver` to the paired `Recorder`

`Recorder` has an async `record` method that streams the input events it
receives from the `Listener` to a yaml file as it receives them.